requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
ccxt>=4.0.0
requests>=2.28.0
flask>=2.3.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For enhanced data analysis
scipy>=1.9.0
scikit-learn>=1.1.0
